backend/requirements.txt
# Backend Dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.5
aiofiles>=23.0.0

# ML/AI
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
transformers>=4.30.0

# Video Processing
opencv-python-headless>=4.8.0
numpy>=1.24.0
scipy>=1.11.0
Pillow>=10.0.0

# Face Detection
facenet-pytorch>=2.5.0

# Utilities
pydantic>=2.0.0
tqdm>=4.65.0

